+++ src/components/Projects.jsx
@@ -1,16 +1,29 @@
 import React from "react";
-import { data } from "../data/works.js";
+import { data } from "../data/projects.js";
 
-const Work = () => {
+const Projects = () => {
   const project = data;
   return (
-    <div name="work" className="w-full md:h-screen text-[#f2f3f5] bg-[#0a172f]">
+    <div
+      name="projects"
+      className="w-full md:h-screen text-[#f2f3f5] bg-[#0a172f]"
+    >
       <div className="max-w-[1000px] mx-auto p-4 flex flex-col justify-center w-full h-full">
         <div className="pb-8">
           <p className="text-3xl font-bold inline border-b-4 border-white">
-            Work
+            Projects
           </p>
-          <p className="py-3">Presenting my recent projects and code</p>
+          <p className="py-3">
+            Presenting my recent projects and code. Check out more on my{" "}
+            <a
+              className="text-[#66CCFF] hover:text-[#f2f3f5]"
+              href="https://github.com/antonrosh"
+              target="_blank"
+              rel="noreferrer"
+            >
+              GitHub
+            </a>
+          </p>
         </div>
 
         {/* Container for projects */}
@@ -49,4 +62,4 @@ const Work = () => {
   );
 };
 
-export default Work;
+export default Projects;
